refactor: move userinfo to header
