+++ web/src/components/DropdownMenu/index.css
@@ -1,9 +1,3 @@
-.dropdown-list {
-    position: absolute;
-    z-index: 1;
-    padding: 0.5rem;
-}
-
 .dropdown-list.hidden {
     display: none;
 }
