+++ web/src/requests.js
@@ -1,3 +1,5 @@
+import { getCookie } from "commons";
+
 /**
  * Create a conversation
  * @returns 
@@ -67,4 +69,14 @@ export const summarizeConversation = async (conversationId) => {
     return fetch(`/api/conversations/${conversationId}/summarization`, {
         method: "POST",
     }).then(res => res.json());
+}
+
+export const logout = async () => {
+    const sessionId = getCookie("authservice_session");
+    await fetch("/authservice/logout", {
+        method: "POST",
+        headers: {
+            Authorization: `Bearer ${sessionId}`,
+        },
+    });
 }
